backend/pyproject.toml
[project]
name = "ragify-backend"
version = "0.1.0"
description = "Backend RAG pour assistant IA Shopify"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
